requirements.txt
# =============================================================================
# NFL Draft Prospect Intelligence System — Dependencies
# =============================================================================

# --- Data Ingestion ---
nflreadpy>=0.1.0
cfbd>=4.0.0
requests>=2.31.0
pandas>=2.1.0
polars>=0.20.0

# --- Feature Engineering ---
numpy>=1.24.0
scipy>=1.11.0

# --- ML / Models ---
scikit-learn>=1.3.0
xgboost>=2.0.0
optuna>=3.4.0
shap>=0.43.0
lifelines>=0.28.0

# --- Visualization ---
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# --- Streamlit App ---
streamlit>=1.29.0

# --- Utilities ---
python-dotenv>=1.0.0
tqdm>=4.66.0

# --- Notebooks ---
jupyter>=1.0.0
nbformat>=5.9.0
ipykernel>=6.25.0
